feat: 🎸 error case
